refactor(v0.72.0): tool_executor.py 1047L → 5-module 패키지로 분할 (Tier 3 #4)
Tier 3 God Object 분할 시리즈 네 번째. ToolExecutor (380L) +
ToolCallProcessor (540L) + 4개 module-level 헬퍼 + 1 spinner
contextmanager가 한 파일에 묶여 있던 1047 LOC monolith를 관심사별
5개 모듈로 분리.

새 패키지 구조 — core/agent/tool_executor/:
  __init__.py    42L  13개 심볼 re-export
  _spinner.py    37L  _tool_spinner contextmanager (lazy console resolve)
  _helpers.py    60L  _compute_model_tool_limit, _guard_tool_result
  executor.py   416L  ToolExecutor + _write_denial_with_fallback +
                      _tool_spinner shim (test monkeypatch 호환)
  processor.py  568L  ToolCallProcessor 본체 (의도적 단일 보존)

공개 API 변경 0건. 외부 import 25곳 (core/ 5, tests/ 19, scripts/ 1)
수정 불필요 — 패키지 __init__.py 가 ToolExecutor, ToolCallProcessor,
_tool_spinner, _compute_model_tool_limit, _guard_tool_result,
_write_denial_with_fallback, console + 6개 module-level 상수 모두 re-export.

pyproject.toml lint-imports 4줄 path rename:
core.agent.tool_executor → core.agent.tool_executor.executor
(import-linter는 leaf 모듈 path 기준이라 mechanical 수정 필요)

Quality Gates:
- ruff/ruff-format/mypy 0 errors (283 source files, +4)
- pytest 4344 passed, 1 skipped, 24 deselected — 변동 없음
- E2E Cowboy Bebop A (68.4) 변동 없음
- 최대 단일 파일 1047 → 568 LOC (-46%)

남은 Tier 3 후보 5개 — commands.py(2441), cli/__init__.py(1889),
agent/loop.py(1754), agentic_ui.py(1160), llm/router.py(1046).
