+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "geode"
-version = "0.71.0"
+version = "0.72.0"
 description = "GEODE — 범용 자율 실행 에이전트"
 license = "Apache-2.0"
 readme = "README.md"
@@ -101,8 +101,8 @@ ignore_imports = [
     "core.agent.loop -> core.cli.session_checkpoint",
     "core.agent.loop -> core.cli.transcript",
     "core.agent.approval -> core.cli",
-    "core.agent.tool_executor -> core.cli.bash_tool",
-    "core.agent.tool_executor -> core.cli.redaction",
+    "core.agent.tool_executor.executor -> core.cli.bash_tool",
+    "core.agent.tool_executor.executor -> core.cli.redaction",
     "core.agent.worker -> core.cli.tool_handlers",
 ]
 
@@ -125,8 +125,8 @@ ignore_imports = [
     "core.agent.worker -> core.cli.tool_handlers",
     "core.agent.loop -> core.cli.session_checkpoint",
     "core.agent.loop -> core.cli.transcript",
-    "core.agent.tool_executor -> core.cli.redaction",
-    "core.agent.tool_executor -> core.cli.bash_tool",
+    "core.agent.tool_executor.executor -> core.cli.redaction",
+    "core.agent.tool_executor.executor -> core.cli.bash_tool",
     "core.agent.approval -> core.cli",
     "core.server.ipc_server.poller -> core.cli",
     "core.server.ipc_server.poller -> core.cli.session_state",
